CDAP-8037 setting to control whether CDAP manages coprocessors
When the setting is set to false, no path to the coprocessor
will be set when creating HBase tables. In this mode, it is
assumed cluster admins will install the CDAP coprocessors on
all regionservers and ensure that they are present on the
regionserver classpath. Doing this would allow that cluster to
perform an HBase rolling upgrade without bringing CDAP down.